Fix notifyCount variable reset problem
Problem Description: When the NotificationsPanel is displayed, the notifyCount variable is set to zero, then the item in the notify list is always increasing, but the notifyCount variable is also incremented. When the NotificationsPanel is hidden, the notify corresponding to the notifyCount variable has already been read.

Solution: Set the notifyCount variable to zero when hiding the NotificationsPanel.