Drop EXPOSE from Dockerfile
There doesn't seem to be a good way to make this respect a param or
envvar passed in, and it's not strictly needed if you're mapping ports.

Having it be present & incorrect in some cases is worse than having it
be absent in some cases where it would be correct. So just drop it
entirely.

Signed-off-by: Jacob Kopczynski <jacob.kopczynski@luthersystems.com>